Refactor client for better data hiding via ES6.
Things got a bit more functional as well. The function exported from the
module is no longer a constructor, but a function that returns a
promise, eliminating the need for the user to explicitly call `init()`.
This is potentially awkward, since the only way to get a handle on the
client now is from `client().then((client) => {});` which may not be
ideal.

All of the "private" functions on the client are now hidden as
properties stored in a `WeakMap`, and the user-provided `baseUrl` is now
read only.